Add everything that we're going to use in the final project, users related
